取得したドキュメントをリストに表示

+++ pages/index.tsx
@@ -19,8 +19,7 @@ import { getAuth } from 'firebase/auth';
 import { NextRouter, useRouter } from 'next/router';
 import { Header } from '@/components/Header';
 import { SubmitHandler, useForm } from 'react-hook-form';
-
-import { collection, doc, setDoc, getDocs, QuerySnapshot, deleteDoc, updateDoc, DocumentData } from 'firebase/firestore';
+import { collection, doc, setDoc, getDocs, QuerySnapshot, deleteDoc, updateDoc } from 'firebase/firestore';
 
 import { v4 as uuidv4 } from 'uuid';
 import { useState, useCallback, useContext, useEffect } from 'react';
@@ -29,9 +28,11 @@ export default function Home() {
 
   const auth = useAuth();
   const currentUser = auth.currentUser;
-
-
+  
   const user = useContext(AuthContext);
+
+  console.log(user);
+
   const router: NextRouter = useRouter();
   const { logout } = useLogout(router);
 
@@ -48,7 +49,6 @@ export default function Home() {
 
   {/* todoにuidをつける */}
   const todoId = uuidv4();
-
   const [todos, setTodos] = useState<DocumentData[]>([]);
 
   {/* todosコレクションの中のドキュメントにはuidを設定してtodoを追加していく*/}
@@ -61,7 +61,7 @@ export default function Home() {
   //第二引数の配列が変化した時のみ実行される。
   // const getTodos = useCallback(() => {
 
-  //   if (!user) return
+  //   if (!user) return 
   //   // 即時関数。宣言と実行を同時に行う。usecallbackの中に即時関数を描き、非同期処理を行う。
   //   // reactHooksと同時に非同期処理ができない。
 
@@ -73,6 +73,7 @@ export default function Home() {
   //   })
   //   console.log('test')
   // }, [todos])
+
   // const getTodos = () => {
   //   const [todos, setTodos] = useState<Todo[]>([
   //     { title: '', status: '' },
@@ -92,13 +93,10 @@ export default function Home() {
   // getTodos();
 
   // console.log(todos);
-
-    // 即時関数。宣言と実行を同時に行う。usecallbackの中に即時関数を描き、非同期処理を行う。
-    // reactHooksと同時に非同期処理ができない。
   useEffect(() => {
     (async () => {
-      if (!currentUser) return
-      const docRef = collection(db, "users", currentUser.uid, "todos");
+      if (!auth.currentUser) return
+      const docRef = collection(db, "users", auth.currentUser.uid, "todos");
       const docSnap = await getDocs(docRef);
       const todosData = docSnap.docs.map((doc) => doc.data())
       setTodos(todosData);
@@ -113,12 +111,14 @@ export default function Home() {
 
   {/* firestoreに保存されている特定のドキュメントを削除 */}
   const deleteTodo = async(todoId) => {
+    if(!currentUser) return
     const todoRef = doc(db, 'users', currentUser.uid, 'todos', todoId);
     await deleteDoc(todoRef);
   }
 
   {/* firestoreに保存されている特定のドキュメントを編集 */}
   const editTodo = async({todoId, title, status}) => {
+    if(!currentUser) return
     const todoRef = doc(db, 'users', currentUser.uid, 'todos', todoId);
     await updateDoc(todoRef, {
       title: title, 
@@ -166,11 +166,13 @@ export default function Home() {
       </Box>
 
       {/* Todoリスト */}
+
       <UnorderedList listStyleType='none'>
+      {todos.map((todo) => (
         <ListItem p={4}>
           <Flex minWidth='max-content' alignItems='center' gap='2'>
             <Box p='2'>
-              <Heading size='md'>Chakra App</Heading>
+              <Heading size='md'>{todo.title}</Heading>
             </Box>
             <Spacer />
             <ButtonGroup gap='2'>
@@ -180,32 +182,7 @@ export default function Home() {
           </Flex>
           <Divider orientation='horizontal' mt='4'/>
         </ListItem>
-        <ListItem p={4}>
-          <Flex minWidth='max-content' alignItems='center' gap='2'>
-            <Box p='2'>
-              <Heading size='md'>Chakra App</Heading>
-            </Box>
-            <Spacer />
-            <ButtonGroup gap='2'>
-              <Button colorScheme='red' onClick={deleteTodo}>削除</Button>
-              <Button colorScheme='blue' onClick={editTodo}>編集</Button>            
-            </ButtonGroup>
-          </Flex>
-          <Divider orientation='horizontal' mt='4'/>
-        </ListItem>
-        <ListItem p={4}>
-          <Flex minWidth='max-content' alignItems='center' gap='2'>
-            <Box p='2'>
-              <Heading size='md'>Chakra App</Heading>
-            </Box>
-            <Spacer />
-            <ButtonGroup gap='2'>
-              <Button colorScheme='red' onClick={deleteTodo}>削除</Button>
-              <Button colorScheme='blue' onClick={editTodo}>編集</Button>            
-            </ButtonGroup>
-          </Flex>
-          <Divider orientation='horizontal' mt='4'/>
-        </ListItem>
+            ))}
       </UnorderedList>
     </>
   );
